Sync Fork from Upstream Repo

+++ src/book/init.rs
@@ -1,12 +1,11 @@
 use std::fs::{self, File};
-use std::path::PathBuf;
 use std::io::Write;
-use toml;
+use std::path::PathBuf;
 
-use config::Config;
 use super::MDBook;
-use theme;
-use errors::*;
+use crate::config::Config;
+use crate::errors::*;
+use crate::theme;
 
 /// A helper for setting up a new book and its directory structure.
 #[derive(Debug, Clone, PartialEq)]
@@ -110,7 +109,8 @@ impl BookBuilder {
     fn copy_across_theme(&self) -> Result<()> {
         debug!("Copying theme");
 
-        let themedir = self.config
+        let themedir = self
+            .config
             .html_config()
             .and_then(|html| html.theme)
             .unwrap_or_else(|| self.config.book.src.join("theme"));
@@ -127,8 +127,20 @@ impl BookBuilder {
         let mut index = File::create(themedir.join("index.hbs"))?;
         index.write_all(theme::INDEX)?;
 
-        let mut css = File::create(themedir.join("book.css"))?;
-        css.write_all(theme::CSS)?;
+        let cssdir = themedir.join("css");
+        fs::create_dir(&cssdir)?;
+
+        let mut general_css = File::create(cssdir.join("general.css"))?;
+        general_css.write_all(theme::GENERAL_CSS)?;
+
+        let mut chrome_css = File::create(cssdir.join("chrome.css"))?;
+        chrome_css.write_all(theme::CHROME_CSS)?;
+
+        let mut print_css = File::create(cssdir.join("print.css"))?;
+        print_css.write_all(theme::PRINT_CSS)?;
+
+        let mut variables_css = File::create(cssdir.join("variables.css"))?;
+        variables_css.write_all(theme::VARIABLES_CSS)?;
 
         let mut favicon = File::create(themedir.join("favicon.png"))?;
         favicon.write_all(theme::FAVICON)?;
@@ -160,15 +172,19 @@ impl BookBuilder {
         let src_dir = self.root.join(&self.config.book.src);
 
         let summary = src_dir.join("SUMMARY.md");
-        let mut f = File::create(&summary).chain_err(|| "Unable to create SUMMARY.md")?;
-        writeln!(f, "# Summary")?;
-        writeln!(f, "")?;
-        writeln!(f, "- [Chapter 1](./chapter_1.md)")?;
+        if !summary.exists() {
+            trace!("No summary found creating stub summary and chapter_1.md.");
+            let mut f = File::create(&summary).chain_err(|| "Unable to create SUMMARY.md")?;
+            writeln!(f, "# Summary")?;
+            writeln!(f)?;
+            writeln!(f, "- [Chapter 1](./chapter_1.md)")?;
 
-        let chapter_1 = src_dir.join("chapter_1.md");
-        let mut f = File::create(&chapter_1).chain_err(|| "Unable to create chapter_1.md")?;
-        writeln!(f, "# Chapter 1")?;
-
+            let chapter_1 = src_dir.join("chapter_1.md");
+            let mut f = File::create(&chapter_1).chain_err(|| "Unable to create chapter_1.md")?;
+            writeln!(f, "# Chapter 1")?;
+        } else {
+            trace!("Existing summary found, no need to create stub files.");
+        }
         Ok(())
     }
